LayoutManager & MasterHandler improvements. Furthermore added support of serializable objects.

+++ src/de/mixedfx/gui/LayoutManager.java
@@ -1,18 +1,23 @@
 package de.mixedfx.gui;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
+
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
 
 import org.apache.commons.io.FileUtils;
 
 import de.mixedfx.file.DataHandler;
 import de.mixedfx.file.FileObject;
+import de.mixedfx.java.StringArrayList;
 import de.mixedfx.logging.Log;
 
 public class LayoutManager
 {
 	public final FileObject				layoutDir;
+
+	public StringProperty				currentLayout;
 
 	public final List<LayoutElement<?>>	layoutElements;
 
@@ -22,6 +27,10 @@ public class LayoutManager
 	 */
 	public LayoutManager(final FileObject layoutDir, final List<LayoutElement<?>> layoutElements, final String defaultLayout)
 	{
+		this.currentLayout = new SimpleStringProperty(defaultLayout);
+		this.layoutDir = layoutDir;
+		this.layoutElements = layoutElements;
+
 		if (!layoutDir.toFile().exists())
 		{
 			try
@@ -33,8 +42,6 @@ public class LayoutManager
 				Log.assets.error("Layout directory could not have been created! " + layoutDir);
 			}
 		}
-		this.layoutDir = layoutDir;
-		this.layoutElements = layoutElements;
 
 		// Apply default layout
 		this.applyLayout(defaultLayout);
@@ -45,7 +52,7 @@ public class LayoutManager
 	 */
 	public List<String> getList()
 	{
-		final ArrayList<String> all = DataHandler.getSubFolderList(this.layoutDir);
+		final StringArrayList all = new StringArrayList(DataHandler.getSubFolderList(this.layoutDir));
 		if (all.size() == 0)
 		{
 			all.add("Default");
@@ -80,10 +87,22 @@ public class LayoutManager
 		{
 			le.update(currentLayout);
 		}
+
+		this.currentLayout.set(layout);
 	}
 
+	/**
+	 * Removes the layout also from the hdd! You can't remove the currentLayout.
+	 *
+	 * @param layout
+	 *            The name of the layout which shall be deleted.
+	 */
 	public void removeLayout(final String layout)
 	{
+		if (layout.equalsIgnoreCase(this.currentLayout.get()))
+		{
+			return;
+		}
 		DataHandler.deleteFile(this.layoutDir.setFullName(layout));
 	}
 }
